# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mystics Audit — Cloud-Based Accounting & GST Compliance SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # CHART OF ACCOUNTS
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}/aging:
    get:
      operationId: getCustomerAging
      tags: [customers]
      summary: Get customer outstanding aging
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Customer aging
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAging"
...
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List budgets
      parameters:
        - { name: fiscalYear, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}/vs-actual:
    get:
      operationId: getBudgetVsActual
      tags: [budgets]
      summary: Budget vs actual comparison
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Budget vs actual report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetVsActual"

  # USERS
...
components:
  schemas:
...
    Company:
      type: object
      properties:
        id: { type: integer }
        legalName: { type: string }
        tradeName: { type: ["string", "null"] }
        pan: { type: ["string", "null"] }
        cin: { type: ["string", "null"] }
        gstin: { type: ["string", "null"] }
        companyType: { type: string }
        industry: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        pincode: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        fiscalYearStart: { type: string }
        gstFilingFrequency: { type: string }
        createdAt: { type: string }
      required: [id, legalName, companyType, fiscalYearStart, gstFilingFrequency, createdAt]
    CompanyInput:
      type: object
      properties:
        legalName: { type: string }
        tradeName: { type: string }
        pan: { type: string }
        cin: { type: string }
        gstin: { type: string }
        companyType: { type: string }
        industry: { type: string }
        phone: { type: string }
        email: { type: string }
        website: { type: string }
        address: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        fiscalYearStart: { type: string }
        gstFilingFrequency: { type: string }
      required: [legalName, companyType]
    CompanyUpdate:
      type: object
      properties:
        legalName: { type: string }
        tradeName: { type: string }
        phone: { type: string }
        email: { type: string }
        website: { type: string }
        address: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        gstFilingFrequency: { type: string }

    # ACCOUNT
...
    Customer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        gstin: { type: ["string", "null"] }
        pan: { type: ["string", "null"] }
        email: { type: string }
        phone: { type: string }
        whatsapp: { type: ["string", "null"] }
        billingAddress: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        pincode: { type: ["string", "null"] }
        creditLimit: { type: number }
        paymentTerms: { type: string }
        currency: { type: string }
        tdsCategory: { type: ["string", "null"] }
        customerGroup: { type: ["string", "null"] }
        openingBalance: { type: number }
        currentBalance: { type: number }
        isActive: { type: boolean }
        createdAt: { type: string }
      required: [id, name, type, email, phone, creditLimit, paymentTerms, currency, openingBalance, currentBalance, isActive, createdAt]
    CustomerInput:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        gstin: { type: string }
        pan: { type: string }
        email: { type: string }
        phone: { type: string }
        whatsapp: { type: string }
        billingAddress: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        creditLimit: { type: number }
        paymentTerms: { type: string }
        currency: { type: string }
        tdsCategory: { type: string }
        customerGroup: { type: string }
        openingBalance: { type: number }
      required: [name, type, email, phone, paymentTerms]
    CustomerUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        billingAddress: { type: string }
        creditLimit: { type: number }
        paymentTerms: { type: string }
        isActive: { type: boolean }
    CustomerAging:
      type: object
      properties:
        customerId: { type: integer }
        customerName: { type: string }
        current: { type: number }
        days0to30: { type: number }
        days31to60: { type: number }
        days61to90: { type: number }
        days91to180: { type: number }
        days180plus: { type: number }
        total: { type: number }
      required: [customerId, customerName, current, days0to30, days31to60, days61to90, days91to180, days180plus, total]
...
    Budget:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        fiscalYear: { type: string }
        type: { type: string }
        totalBudget: { type: number }
        totalActual: { type: number }
        utilizationPct: { type: number }
        status: { type: string }
        createdAt: { type: string }
      required: [id, name, fiscalYear, type, totalBudget, totalActual, utilizationPct, status, createdAt]
    BudgetInput:
      type: object
      properties:
        name: { type: string }
        fiscalYear: { type: string }
        type: { type: string }
        lines: { type: array, items: { $ref: "#/components/schemas/BudgetLineInput" } }
      required: [name, fiscalYear, type, lines]
    BudgetLineInput:
      type: object
      properties:
        accountId: { type: integer }
        department: { type: string }
        annualAmount: { type: number }
        monthlyAmounts: { type: array, items: { type: number } }
      required: [accountId, annualAmount]
    BudgetVsActual:
      type: object
      properties:
        budgetId: { type: integer }
        budgetName: { type: string }
        fiscalYear: { type: string }
        period: { type: ["string", "null"] }
        lines: { type: array, items: { $ref: "#/components/schemas/BudgetVsActualLine" } }
        totalBudget: { type: number }
        totalActual: { type: number }
        totalVariance: { type: number }
      required: [budgetId, budgetName, fiscalYear, lines, totalBudget, totalActual, totalVariance]
    BudgetVsActualLine:
      type: object
      properties:
        accountCode: { type: string }
        accountName: { type: string }
        department: { type: ["string", "null"] }
        annualBudget: { type: number }
        periodBudget: { type: number }
        periodActual: { type: number }
        periodVariance: { type: number }
        ytdBudget: { type: number }
        ytdActual: { type: number }
        ytdVariance: { type: number }
        utilizationPct: { type: number }
        alertLevel: { type: ["string", "null"] }
      required: [accountCode, accountName, annualBudget, periodBudget, periodActual, periodVariance, ytdBudget, ytdActual, ytdVariance, utilizationPct]

    # USER
...